made IGameEventListenable interface, assert added, and duplicate check on gameeventlistener list
