+++ Assets/BossRoom/Scripts/Shared/ScriptableObjects/IGameEventListenable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 099e24cac73964e498d5bf73bc19973d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
