+++ Assets/BossRoom/Scripts/Shared/ScriptableObjects/UnityEventGameEventListener.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b485354c9406ae843a51583d81e67530
+guid: 1ddfeda1c7e74b74d9baaca2a96c95d2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
